--- webtoon-core/src/test/kotlin/com/depromeet/webtoon/core/crawl/naver/NaverCrawlerFetchAdapterTest.kt
@@ -1,10 +1,0 @@
-package com.depromeet.webtoon.core.crawl.naver
-
-import io.kotest.core.spec.style.FunSpec
-
-internal class NaverCrawlerFetchAdapterTest : FunSpec({
-    xtest("crawl naver") {
-        val crawl = NaverCrawlerFetchAdapter().crawl()
-        println(crawl)
-    }
-})
