+++ webtoon-crawler/build.gradle.kts
@@ -1,4 +1,7 @@
 dependencies {
+    api(project(":webtoon-common"))
+    implementation("org.springframework.boot:spring-boot-starter")
+
     implementation("com.squareup.okhttp3:okhttp:4.9.1")
     implementation("org.jsoup:jsoup")
 }
