+++ webtoon-crawler/src/main/kotlin/com/depromeet/webtoon/crawl/crawler/daum/dto/webtoonlist/Artists.kt
@@ -1,0 +1,7 @@
+package com.depromeet.webtoon.crawl.crawler.daum.dto.webtoonlist
+
+data class Artists(
+
+    val name: String,
+    val penName: String
+)
