+++ webtoon-crawler/src/main/kotlin/com/depromeet/webtoon/crawl/crawler/daum/dto/webtoonlist/Cartoon.kt
@@ -1,0 +1,6 @@
+package com.depromeet.webtoon.crawl.crawler.daum.dto.webtoonlist
+
+data class Cartoon(
+
+    val artists: List<Artists>
+)
