+++ webtoon-crawler/src/main/kotlin/com/depromeet/webtoon/crawl/crawler/daum/dto/webtoonlist/CompleteDaumWebtoonCrawlResult.kt
@@ -1,0 +1,12 @@
+package com.depromeet.webtoon.crawl.crawler.daum.dto.webtoonlist
+
+data class CompleteDaumWebtoonCrawlResult(
+
+    val data: List<CompleteData>
+
+)
+
+data class CompleteData(
+    val nickname: String,
+    val ageGrade: Int
+)
