+++ webtoon-crawler/src/main/kotlin/com/depromeet/webtoon/crawl/crawler/daum/dto/webtoonlist/EpisodeImage.kt
@@ -1,0 +1,7 @@
+package com.depromeet.webtoon.crawl.crawler.daum.dto.webtoonlist
+
+data class EpisodeImage(
+
+    val url: String,
+    val serviceStatus: String
+)
