+++ webtoon-crawler/src/main/kotlin/com/depromeet/webtoon/crawl/crawler/daum/dto/webtoonlist/Page.kt
@@ -1,0 +1,9 @@
+package com.depromeet.webtoon.crawl.crawler.daum.dto.webtoonlist
+
+data class Page(
+
+    val size: Int,
+    val no: Int,
+    val groupRange: Int,
+    val totalItemCount: Int
+)
