+++ webtoon-crawler/src/main/kotlin/com/depromeet/webtoon/crawl/crawler/daum/dto/webtoonlist/ThumbnailImage.kt
@@ -1,0 +1,8 @@
+package com.depromeet.webtoon.crawl.crawler.daum.dto.webtoonlist
+
+data class ThumbnailImage(
+
+    val id: Int,
+    val url: String,
+    val serviceStatus: String
+)
