+++ webtoon-crawler/src/main/kotlin/com/depromeet/webtoon/crawl/crawler/daum/dto/webtoonlist/ThumbnailImage2.kt
@@ -1,0 +1,7 @@
+package com.depromeet.webtoon.crawl.crawler.daum.dto.webtoonlist
+
+data class ThumbnailImage2(
+
+    val id: Int,
+    val url: String
+)
